VideoLaVIT/requirements.txt
contexttimer
decord
diffusers==0.25.1
einops
ftfy
iopath
ipython
opencv-python-headless==4.5.5.64
packaging
pandas
plotly
pre-commit
pycocoevalcap
pycocotools
python-magic
scikit-image
sentencepiece
spacy
streamlit
timm==0.4.12
tqdm
transformers==4.33.2
wheel
torchmetrics
tiktoken
jsonlines
motion-vector-extractor